microservices/notification-service/requirements.txt
flask
python-telegram-bot
flask-cors
py-eureka-client
